+++ package.json
@@ -11,13 +11,14 @@
     "coverage": "jest --coverage"
   },
   "dependencies": {
+    "@cloudinary/react": "^1.11.2",
+    "@cloudinary/url-gen": "^1.11.0",
     "@reduxjs/toolkit": "^1.9.5",
     "@types/node": "20.4.1",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
     "autoprefixer": "10.4.14",
     "axios": "^1.4.0",
-    "cloudinary": "^1.39.0",
     "eslint": "8.44.0",
     "eslint-config-next": "13.4.9",
     "mongoose": "^7.4.1",
